v5/requirements.txt
fastapi[all]==0.112.2
numpy==2.1.0
redis==5.0.8
watchdog==5.0.2
redis==5.0.8
python-dotenv==1.0.1
matplotlib==3.9.2
